requirements.txt
# Build tools (install first)
setuptools>=65.0.0
wheel>=0.38.0
pip>=23.0.0

# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0

# Web automation and crawling
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Web framework for UI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Database and state management
sqlalchemy>=2.0.0
redis>=5.0.0

# AI and ML
openai>=1.3.0
anthropic>=0.7.0
numpy>=1.24.0
tiktoken>=0.5.0

# Logging and monitoring
loguru>=0.7.0

# Security and web testing
httpx>=0.25.0

# Utility libraries
pydantic>=2.5.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0